+++ tomcat/src/main/java/org/apache/coyote/http11/controller/HomeController.java
@@ -1,23 +1,25 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.controller;
 
+import org.apache.coyote.http11.service.HttpService;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 
-public class LoginController extends AbstractController {
+public class HomeController extends AbstractController {
 
     private final HttpService httpService;
 
-    public LoginController(final HttpService httpService) {
+    public HomeController(final HttpService httpService) {
         this.httpService = httpService;
     }
 
     @Override
     protected HttpResponse doPost(final HttpRequest request) {
-        return httpService.login(request);
+        // 아직 기능 존재X
+        return null;
     }
 
     @Override
     protected HttpResponse doGet(final HttpRequest request) {
-        return httpService.getLoginPage(request);
+        return httpService.indexProcess(request.getHttpRequestLine());
     }
 }
