+++ tomcat/src/main/java/org/apache/coyote/http11/controller/StaticResourceController.java
@@ -1,6 +1,7 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.controller;
 
 import java.net.URL;
+import org.apache.coyote.http11.response.HttpStatus;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.response.HttpResponseBody;
