+++ tomcat/src/main/java/org/apache/coyote/http11/response/HttpStatus.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.response;
 
 public enum HttpStatus {
     SUCCESS(200, "OK"),
